HDX-6124 remove files from filestore at hdx_dataset_purge
